requirements.txt
flask==2.3.0
flask-cors==4.0.0
flask-sock==0.7.0
psycopg2-binary==2.9.9
requests==2.31.0
gunicorn==21.2.0
werkzeug==2.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
urllib3==2.1.0
